VW-931 Create production process

+++ Infrastructure/EntityFramework/ManufacturingDB/ProductionStep.cs
@@ -13,9 +13,9 @@ namespace VErp.Infrastructure.EF.ManufacturingDB
         public long ProductionStepId { get; set; }
         public int? StepId { get; set; }
         public string Title { get; set; }
-        public int? ParentId { get; set; }
+        public long? ParentId { get; set; }
         public int ContainerTypeId { get; set; }
-        public long ContainerId { get; set; }
+        public int ContainerId { get; set; }
         public DateTime CreatedDatetimeUtc { get; set; }
         public int CreatedByUserId { get; set; }
         public bool IsDeleted { get; set; }
